+++ Test_Project1/Src/locators.py
@@ -1,0 +1,30 @@
+class LoginPageLocators:
+    LOGIN_URL = "https://opensource-demo.orangehrmlive.com/web/index.php/auth/login"
+    USERNAME_INPUT = "//input[@name='username']"
+    PASSWORD_INPUT = "//input[@name='password']"
+    SUBMIT_BUTTON = "//button[@type='submit']"
+    PROFILE_BUTTON = "//span[@class='oxd-userdropdown-tab']"
+    LOGOUT_BUTTON = "//a[text()='Logout']"
+    VALID_USERNAME = "Admin"
+    VALID_PASSWORD = "admin123"
+    INVALID_USERNAME = "admin"
+    INVALID_PASSWORD = "boom"
+
+
+class PIMPageLocators:
+    PIM_MODULE = "//span[text()='PIM']"
+    ADD_EMPLOYEE_BUTTON = "//a[text()='Add Employee']"
+    FIRST_NAME_INPUT = "//input[@name='firstName']"
+    LAST_NAME_INPUT = "//input[@name='lastName']"
+    SAVE_BUTTON = "//button[text()=' Save ']"
+    SUCCESS_MESSAGE = "//p[text()='Successfully Saved']"
+    EMPLOYEE_LIST_BUTTON = "//a[text()='Employee List']"
+    SELECT_USER_DELETE = "((//div[@role='row'])[3]//child::div)[1]"
+    DELETE_BUTTON = "//button[text()=' Delete Selected ']"
+    CONFIRM_DELETE_BUTTON = "//button[text()=' Yes, Delete ']"
+    DELETE_SUCCESS_MESSAGE = "//p[text()='Successfully Deleted']"
+    SELECT_USER_EDIT = "((//div[@role='row'])[3]//child::div)[1]"
+    MALE_BUTTON = '//*[@id="app"]/div[1]/div[2]/div[2]/div/div/div/div[2]/div[1]/form/div[3]/div[2]/div[2]/div/div[2]/div[1]/div[2]/div/label/span'
+    SUBMIT_EDIT_BUTTON = "(//button[@type='submit'])[1]"
+    UPDATE_SUCCESS_MESSAGE = "//p[text()='Successfully Updated']"
+    EDIT_BUTTON = "((//div[@role='row'])[3]//child::div)[19]"
